requirements.txt
allure-pytest==2.11.1
allure-python-commons==2.11.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
et-xmlfile==1.1.0
exceptiongroup==1.2.2
Faker==26.0.0
idna==3.7
iniconfig==2.0.0
jsonpath==0.82.2
loguru==0.7.2
openpyxl==3.1.5
packaging==24.1
pluggy==1.5.0
PyMySQL==1.1.1
pytest==8.0.2
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
tomli==2.0.1
urllib3==2.2.2
win32-setctime==1.1.0
